Hoist the 16 identical strategy methods into MapStrategy
maplibre.strategy and mapbox.strategy each carried their own copy of
fitBounds, flyTo, jumpTo, resize, zoomIn/Out, resetNorth, removeSource,
setSourceData, setMap3dTerrain, the trip methods and unsetDirections — all
byte-identical, and all previously shadowing empty stubs on the base.

They move to MapStrategy, which now also owns the trip layerGroups registry
and exposes fitMapToTrips as an overridable hook. getBounds is left duplicated
for now; its multi-line return type makes the move less mechanical.
